nearly all wombat workflows working but maxquant
