+++ proteobench/io/parsing/io_parse_settings/Quant/lfq/ion/DDA/parse_settings_wombat.toml
@@ -25,7 +25,7 @@
 "_HUMAN" = "HUMAN"
 
 [modifications_parser]
-"parse_column" = "Modified Sequence"
+"parse_column" = "Sequence"
 "before_aa" = false
 "isalpha" = true
 "isupper" = true
