+++ proteobench/io/parsing/parse_settings.py
@@ -130,6 +130,7 @@ class ParseSettings:
 
         df.rename(columns=self.mapper, inplace=True)
 
+
         replicate_to_raw = defaultdict(list)
         for k, v in self.condition_mapper.items():
             replicate_to_raw[v].append(k)
