--- webinterface/.streamlit/config.toml
@@ -1,10 +1,0 @@
-[theme]
-base="light"
-#FFFFFF
-
-[global]
-disableWidgetStateDuplicationWarning = true
-
-[server]
-enableStaticServing = true
-maxUploadSize = 1000
