--- webinterface/.streamlit/pages.toml
@@ -1,19 +1,0 @@
-[[pages]]
-path = "Home.py"
-name = "Home"
-
-[[pages]]
-path = "pages/module_pages/quant_lfq_ion_DDA.py"
-name = "DDA Quant Ion Level -BETA-"
-
-[[pages]]
-path = "pages/module_pages/quant_lfq_peptidoform_DDA.py"
-name = "DDA Quant Peptidoform Level -BETA-"
-
-[[pages]]
-path = "pages/module_pages/quant_lfq_ion_DIA_AIF.py"
-name = "DIA Quant Ion Level - AIF -BETA-"
-
-[[pages]]
-path = "pages/module_pages/quant_lfq_ion_DIA_diaPASEF.py"
-name = "DIA Quant Ion Level - diaPASEF -BETA-"
